feat: renaming aws folder to adapters
