+++ apps/notifications/src/adapters/dynamo.ts
