+++ apps/notifications/src/adapters/ses.ts
