+++ apps/notifications/tsconfig.json
@@ -11,7 +11,7 @@
       "outDir": "./dist",
       "baseUrl": ".",
       "paths": {
-        "@/*": ["src/*"],
+        "@/*": ["src/*"]
       },
       "incremental": true,
       "resolveJsonModule": true,
